+++ example/lib/qubits/settings_super_qubit.dart
@@ -2,8 +2,7 @@ import 'package:super_qubit/super_qubit.dart';
 import 'theme_qubit.dart';
 
 class SettingsSuperQubit extends SuperQubit {
-  @override
-  void init() {
+  SettingsSuperQubit() {
     // Listen to theme changes
     listenTo<ThemeQubit>((state) {
       print('Theme changed: ${state.isDark ? "Dark" : "Light"} mode');